CheckRequest is now AnalyzeRequest
The CheckRequest implementation was doing to much, and was concerned
with how the request related to the controller. Now it just takes a
request and acts on it. The logic relevent to checking if a request
should be handled as json that deals with the controller has been moved
back into the View because it decides whether or not to even use the
AnalyzeRequest object.